% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchClust.R
\name{benchClust}
\alias{benchClust}
\title{Benchmark clustering output}
\usage{
benchClust(dat, id_var, t_var, y_var, g_var, arr_cluster, seed = 1)
}
\arguments{
\item{dat}{Data frame of dataset, or list of a few datasets to benchmark}

\item{id_var}{Character of id variable}

\item{t_var}{Column name in \code{dat} representing the scaled time points ranging in \eqn{[0,1]}}

\item{y_var}{Character of column name in \code{dat} as observation}

\item{g_var}{Column name in \code{dat} containing true group labels in the dataset}

\item{arr_cluster}{Array of predicted clusters by projection clustering method,
passed from function \code{\link{postPairs}}}

\item{seed}{Random seed to pass to benchmark methods}
}
\value{
List of matrices with rows indicating the benchmark methods and two columns
Rand Index and adjusted Rand Index, corresponding to the list of \code{dat}
}
\description{
Evaluate clustering result by calculating Rand Index and adjusted Rand
Index, benchmarking against 6 traditional clustering techniques:
(1) HC_dist: hierarchical clustering
based on an inte- grated periodogram-based method as dissimilarity
measure (Montero and Vilar, 2014)
(2) HC_pred: hierarchical clustering based on a prediction
density-based method as dissimilarity measure (Montero and Vilar, 2014)
(3) BHC: Bayesian model-based hierarchical clustering with accounting
for uncertainty using the Dirichlet process (Savage et al., 2009)
(4) Mclust: finite Gaussian mixture model under Bayesian framework
estimated by Estimation-Maximisation (Scrucca et al., 2016)
(5) VC: clustering based on Bayesian mixtures of linear mixed models
estimated via variational inference (Tan and Nott, 2014).
(Omitted) KML: K-means for longitudinal data (Genolini and Falissard, 2011).
The relevant code is commented out but executable under local environment
}
\examples{
benchClust(DATASET,'ID','t','Record','Group',out_pc$arr_cluster)

}
