% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustBoot.R
\name{clustBoot}
\alias{clustBoot}
\title{Choose cluster number by bootstrap method}
\usage{
clustBoot(mat_fitted, dat, nB, nCl, seed)
}
\arguments{
\item{mat_fitted}{Prediction replicates of Linear Mixed Model output. Multiple columns
corresponds to sets of random effects of projection defined in \code{pcFit}}

\item{dat}{Longitudinal data input}

\item{nB}{Number of bootstrap samples}

\item{nCl}{Maximum number of clusters to consider}

\item{seed}{Random seed to pass to bootstrap}
}
\value{
list(sBtb, cutoff, nClusters)
\item{sBtb}{Matrix of clustering distance, with \code{nCl} rows and \code{ncol(mat_fitted} columns
corresponding to projections in \code{mat_fitted}}
\item{cutoff}{Vector of cutoff values: \eqn{cutoff_j=1/2*max(sBtb[,j])}}
\item{nClusters}{Vector of cluster numbers chosen by the minimum number with
clustering distance no larger than the cutoff}
}
\description{
With predicted projections from \code{pcFit}, take bootstrap samples for
\code{nBoot} times, at each time two random samples are drawn and applied in
\code{kmeans} clustering with both input sets to create 4 sets of clustering results.
Empirical clustering distance is hence computed to enable choosing minimal cluster number
with cutoff being half of the largest clustering distance.
}
\examples{
\dontrun{
df_of_draws <- modelStan("Record", paste0("Z", 1:10), "ID", DATASET)
}
ls_par <- postMean(df_of_draws, paste0("Z", 1:10), "ID", DATASET)
}
\seealso{
Other Cluster number choices: 
\code{\link{clustKL}()}
}
\concept{Cluster number choices}
