% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustKL.R
\name{clustKL}
\alias{clustKL}
\title{Choose cluster number by KL method}
\usage{
clustKL(id_var, ls_par, dat, ls_idxA, nIter, thKL, regQ, seed)
}
\arguments{
\item{id_var}{Character of id variable}

\item{ls_par}{Posterior mean of LMM parameters generated from \link{postMean} output}

\item{dat}{Longitudinal data input}

\item{ls_idxA}{List of random effect indices to project on}

\item{nIter}{Number of iterations used in clustering optimization}

\item{thKL}{Proportion of max KL, used as upper threshold of Kullback-Leibler divergence corresponding to chosen number of clusters}

\item{regQ}{Positive regularization value to add to the diagonal of matrix to be inverted}

\item{seed}{Random seed to initialize cluster centers}
}
\value{
list(nClusters,KLs,cluster0)
\item{nClusters}{Vector of optimised cluster numbers corresponding to sets of random effects in \code{ls_idxA}}
\item{KLs}{List of Kullback-Leibler divergence across grid of cluster numbers corresponding to \code{ls_idxA}}
\item{cluster0}{List of matrix, column k indicates cluster membership for k clusters corresponding to \code{ls_idxA}}
}
\description{
To find minimal cluster number which reduces Kullback-Leibler divergence
no larger than a customized threshold. KL monotonically decreases with number
of clusters.
}
\examples{
data(df_of_draws)
ls_par <- postMean(df_of_draws, paste0("Z", 1:10), "ID", DATASET)
ls_idxA <- list(
  seq(10),
  1:4,
  5:7,
  8:10
)
out_KL <- clustKL("ID", ls_par, DATASET, ls_idxA, 10, .1, 1e-6, 1)
}
\seealso{
Other BayesPC main functions: 
\code{\link{modelStan}()},
\code{\link{pcFit}()},
\code{\link{postMean}()}

Other Cluster number choices: 
\code{\link{clustBoot}()}
}
\concept{BayesPC main functions}
\concept{Cluster number choices}
