% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_of_draws.R
\docType{data}
\name{df_of_draws}
\alias{df_of_draws}
\title{Sample \code{modelStan} output to demonstrate use of package.}
\format{
A dataframe of 10 rows and 7524 columns:
\describe{
\item{tau}{MCMC output for parameter tau based on stan script}
...
}
}
\usage{
df_of_draws
}
\description{
First 100 rows of posterior output generated by using \code{modelStan} function based on \code{DATASET}.
}
\keyword{datasets}
