% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStan.R
\name{modelStan}
\alias{modelStan}
\title{Bayesian mixed linear regression modelling}
\usage{
modelStan(y_var, x_var, id_var, dat, ...)
}
\arguments{
\item{y_var}{character of response variable}

\item{x_var}{character vector of random effect variables}

\item{id_var}{character of id variable}

\item{dat}{dataset to model with stan}

\item{...}{Additional arguments passed to \link[rstan]{sampling}}
}
\value{
dataframe of posterior output from Bayesian mixed linear regression
}
\description{
To fit longitudinal dataset using linear mixed model by calling stan package.
}
\examples{
data(DATASET)
df_of_draws <- modelStan("Record", paste0("Z", 1:10), "ID", DATASET, iter = 2, chain = 1)
}
\seealso{
Other BayesPC main functions: 
\code{\link{pcFit}()},
\code{\link{postMean}()}
}
\concept{BayesPC main functions}
