% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_pc.R
\docType{data}
\name{out_pc}
\alias{out_pc}
\title{Sample \code{\link{postPairs}} output to demonstrate use of package.}
\format{
list
}
\usage{
out_pc
}
\description{
List of 2 elements.
\enumerate{
\item A list containing matrix of probabilities respective to random effect sets,
with rows and columns corresponding to subject ID. Cell values lie between 0 and 1.
Co-membership probability table is based on 1000 posterior draws based on output
from \code{\link{postPairs}} function using the simulated dataset \code{DATASET},
accessible by \code{data(DATASET)}.
\item Array of cluster labels based on the posterior draws, with each column
corresponding to given random effects.
}
}
\keyword{datasets}
