% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcFit.R
\name{pcFit}
\alias{pcFit}
\title{Projection clustering function}
\usage{
pcFit(id_var, ls_par, dat, ls_idxA)
}
\arguments{
\item{id_var}{Character of id variable}

\item{ls_par}{Posterior mean of LMM parameters generated from \link{postMean} output}

\item{dat}{Longitudinal data input}

\item{ls_idxA}{List of random effect indices to project on}
}
\value{
Matrix containing fitted values, with each column
corresponding to a set of chosen random effects indices
specified in \code{ls_idxA}
}
\description{
Compute LMM fitted values based on posterior mean output of parameters,
with chosen random effects to project on.
}
\examples{
\dontrun{
df_of_draws <- modelStan("Record", paste0("Z", 1:10), "ID", DATASET)
}
ls_par <- postMean(df_of_draws, paste0("Z", 1:10), "ID", DATASET)
ls_idxA <- list(
  seq(10),
  1:4,
  5:7,
  8:10
)
mat_fitted <- pcFit("ID", ls_par, DATASET, ls_idxA)
}
\seealso{
Other BayesPC main functions: 
\code{\link{clustKL}()},
\code{\link{modelStan}()},
\code{\link{postMean}()}
}
\concept{BayesPC main functions}
