% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairs.R
\name{plotPairs}
\alias{plotPairs}
\title{Plot pairwise probability of co-membership}
\usage{
plotPairs(
  dat,
  id_var,
  g_var,
  ls_prob,
  ls_groups = NULL,
  ls_labels = NULL,
  cutPoints = c(0.5, 0.8),
  subTitles = paste(c("All", "Low", "Medium", "High"), "frequencies"),
  allColors = c("darkseagreen", "coral3", "steelblue", "sandybrown"),
  seed = 1,
  thin = 0.1
)
}
\arguments{
\item{dat}{Longitudinal data input}

\item{id_var}{Character of id variable}

\item{g_var}{Column name in \code{dat} containing true group labels in the dataset}

\item{ls_prob}{List of pairwise probability output of function \code{\link{postPairs}}}

\item{ls_groups}{List of "similar" groups as elements of list respective to groups
as list names. The co-membership probabilities indicated "similar" groups will be
plotted in color \code{allColors},
and "dissimilar" groups will be plotted in black}

\item{ls_labels}{List of group labels to display sequentially on the plot. Default is
alphabetically sorted groups}

\item{cutPoints}{Vector of two cutoff points between 0 and 1 to
discretize probabilities into three intervals to indicate three levels of transparency
of the curves. Default c(.5,.8)}

\item{subTitles}{Vector of titles to be labeled for each subplot of random effect scenario}

\item{allColors}{Colors of the curves under the random effect scenarios. Default
c('darkseagreen', 'coral3','steelblue','sandybrown')}

\item{seed}{Random seed for sampling curves when \code{thin} is less than 1}

\item{thin}{Thinning parameter between 0 and 1. Recommended to set at low value to
aid visualization. When \code{thin}=1 all curves are shown. Default is 0.1}
}
\description{
Given co-membership probability input based on \code{\link{postPairs}},
create a linkage plot colored according to probability range.
}
\examples{
ls_prob <- out_pc$ls_prob
ls_labels <- rep(list(1:4),4)
ls_labels[[4]] <- c(1,3,2,4)
ls_groups <- list(
'1'= c('2','3'),
'2'= c('1','4'),
'3'= c('1','4'),
'4'= c('3','2')
)
plotPairs(DATASET,'ID','Group',ls_prob,ls_groups,ls_labels,thin=.01)
}
