% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postPairs.R
\name{postPairs}
\alias{postPairs}
\title{Calculate pairwise co-membership probability}
\usage{
postPairs(
  df_of_draws,
  x_var,
  id_var,
  dat,
  ls_idxA,
  nIter,
  nDraw,
  nClusters,
  regQ,
  seed
)
}
\arguments{
\item{df_of_draws}{Data frame of simulated LMM output}

\item{x_var}{Character vector of random effect variables}

\item{id_var}{Character of id variable}

\item{dat}{Longitudinal data input}

\item{ls_idxA}{List of random effect indices to project on}

\item{nIter}{Number of iterations used in clustering optimization}

\item{nDraw}{Number of draws to sample from projection clustering output}

\item{nClusters}{Vector of cluster numbers}

\item{regQ}{Positive regularization value to add to the diagonal of matrix to be inverted}

\item{seed}{Random seed to initialize cluster centers}
}
\value{
List of pairwise probability tables corresponding to random effect projection
specified in \code{ls_idxA}. Row and column index of table indicate subject
ID number in \code{dat}
}
\description{
Given a vector of cluster numbers and projection clustering output, \code{postPairs}
calculate the posterior probability of any pair of subjects being clustered in the
same group.
}
\examples{
\dontrun{
df_of_draws <- modelStan("Record", paste0("Z", 1:10), "ID", DATASET)
}
ls_idxA <- list(
  seq(10),
  1:4,
  5:7,
  8:10
)
ls_prob <- postPairs(df_of_draws, x_var=paste0("Z", 1:10), id_var="ID", dat=DATASET,
                    ls_idxA, nIter=10, nDraw=2, nClusters=4, regQ=1e-6, seed=1)
}
